add synchronize block
This adds a Synchronize block. The block has two inputs and one
output. It consumes data from both inputs simultaneously and passes
the samples from the first input to the output. The intended application
of this block is to apply backpressure to a transmitter that is to
be connected to a DIFI Sink, using the output from a DIFI Source to
control the rate at which samples flow through the Synchronize block.