difi_source: add message port to output time tags
This adds an optional message port that outputs a message with
the pck_n dictionary (contains the timestamp and the packet sequence
number) every time that such a dictionary is inserted as a tag in
the output IQ stream (this happens at the beginning and whenever
there is packet loss).

This feature can be used to update the time of other blocks that
are not connected downstream of the DIFI source (so they don't receive
the stream tags).