use Volk to pack/unpack IQ samples
This modifies the DIFI Sink and Source so that they use Volk to
pack/unpack the IQ samples, greatly improving the performance (CPU usage
down by more than a factor of 2 in some benchmarks).

A new template parameter S is added to the sink and source blocks to
indicate if DIFI packets use std::complex<int8_t> or
std::complex<int16_t>. Therefore, 4 variants of each sink/source block
are generated, allowing us to choose the appropriate Volk kernel for
the IQ sample conversion at compile time.

The scaling in the DIFI Sink has been simplified. Instead of supporting
different scaling modes, the input is converted from the range [-1.0,
1.0] to [-127, 127] or [-32767, 32767]. If custom scaling is required,
it can be accomplished with Add Const and Multiply Const blocks outside
of the DIFI Sink.

Additionally, instead of allocating a std::vec to contain the DIFI
packet in the DIFI Sink every time that a packet is sent, the std::vec
is allocated in the class constructor and re-used every time.